pyproject.toml
[tool.poetry]
name = "policy-rag"
version = "0.1.0"
description = "Policy document RAG system"
authors = ["rianto42 <rianto.utomo@student.uts.edu.au>"]
readme = "README.md"
packages = [{include = "policy_rag", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pymupdf = ">=1.26.0,<2.0.0"
transformers = ">=4.52.4,<5.0.0"
torch = ">=2.7.1,<3.0.0"
matplotlib = ">=3.10.3,<4.0.0"
scikit-learn = ">=1.7.0,<2.0.0"
chromadb = ">=1.0.12,<2.0.0"
numpy = ">=1.26.4,<2.0.0"
sentence-transformers = ">=2.5.1,<3.0.0"
streamlit = "^1.45.1"
huggingface-hub = "^0.33.0"
accelerate = "^1.7.0"
sentencepiece = "^0.2.0"
langchain = "^1.2.0"
langchain-google-genai = "^4.1.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
